modify code to create compatibility with mysql
