+++ db/mysql/rollback/patch-initial-schema.sql
@@ -1,0 +1,25 @@
+drop table if exists heartbeat;
+
+drop table if exists incident;
+
+drop table if exists monitor_group;
+
+drop table if exists `group`;
+
+drop table if exists monitor_notification;
+
+drop table if exists monitor_tag;
+
+drop table if exists monitor_tls_info;
+
+drop table if exists notification;
+
+drop table if exists notification_sent_history;
+
+drop table if exists monitor;
+
+drop table if exists setting;
+
+drop table if exists tag;
+
+drop table if exists user;
