+++ package.json
@@ -79,6 +79,7 @@
         "iconv-lite": "^0.6.3",
         "jsonwebtoken": "~8.5.1",
         "jwt-decode": "^3.1.2",
+        "knex": "^0.95.14",
         "limiter": "^2.1.0",
         "nodemailer": "~6.6.5",
         "notp": "~2.0.3",
